Move QuickHull3D from geometric3D namespace to descriptor3D namespace.
