+++ src/main/java/net/imagej/ops/descriptor3d/QuickHull3DFromMC.java
@@ -5,7 +5,7 @@ import org.scijava.plugin.Plugin;
 import net.imagej.ops.AbstractFunctionOp;
 import net.imagej.ops.FunctionOp;
 import net.imagej.ops.Op;
-import net.imagej.ops.Ops.Geometric3D.ConvexHull3D;
+import net.imagej.ops.Ops.Descriptor3D.ConvexHull3D;
 import net.imglib2.roi.IterableRegion;
 import net.imglib2.type.BooleanType;
